+++ src/main/java/com/fastcampus/reserve/domain/user/UserDetailsServiceImpl.java
@@ -32,7 +32,9 @@ public class UserDetailsServiceImpl implements UserDetailsService {
         );
     }
 
-    private Collection<? extends GrantedAuthority> convertToAuthorities(List<Authority> authorities) {
+    private Collection<? extends GrantedAuthority> convertToAuthorities(
+        List<Authority> authorities
+    ) {
         return authorities.stream()
                 .map(authority -> new SimpleGrantedAuthority(authority.getRole().name()))
                 .collect(Collectors.toList());
